Gutted and reorganized files
