--- app/templates/app/charts/planet_system.js
@@ -1,111 +1,0 @@
-// note that the clickhandler for the solar
-var p_objectColors = {
-    'G': '#FDB813',
-    'moon':'#F4F1C9',
-    'terrestrial':'#3644E4',
-    'ice':'#A7DEDA',
-    'dwarf':'#0EC0A6',
-    'gas':'#0EC0A6'
-}
-
-// dict of the value you want shown and `label` that you want as the column label
-var planet_table_lables = [{"label":"Name","value":"name"}, 
-                            {"label":"Class","value":"class"},
-                            {"label":"Mass","value":"mass"},
-                            {"label":"Radius","value":"radius"},
-                            {"label":"Orbiting Distance","value":"orbitsDistance"},
-                            {"label":"Orbiting Object","value":"orbitsName"},
-                            {"label":"Supports Life","value":"isSupportsLife"}
-                        ]
-
-var table_lables = [{"label":"Name","value":"name"}, 
-                            {"label":"Aggression","value":"aggression"},
-                            {"label":"Conformity","value":"conformity"},
-                            {"label":"Constitution","value":"constitution"},
-                            {"label":"Literacy","value":"literacy"},
-                            {"label":"Idle","value":"isIdle"}
-                        ]
-     
-var desire_table_lables = [{"Type":"Name","value":"type"}, 
-                        {"label":"Name","value":"name"},
-                        {"label":"Weight","value":"weight"},
-                        {"label":"type","value":"type"},
-                        {"label":"comment","value":"comment"},
-                        {"label":"leadingAttribute","value":"leadingAttribute"}
-                    ]
-
-
-var faction_table_lables = [{"label":"Faction Name","value":"name"}]
-
-function clickTablePlanet(d){
-    $.ajax({
-        url: '/ajax/planet-inhabitants',
-        type: 'get',
-        data: d,
-        dataType: 'json',
-        beforeSend: function () {
-            d3.selectAll('#peopleTable').remove()
-            d3.selectAll('#factionTable').remove()
-            d3.selectAll('#description').remove()
-            d3.selectAll('#peopledesires').remove()
-        },
-        success: function(data){
-            console.log(data)
-            if ("pops" in data){
-                draw_table(
-                    "peopleTable",
-                    data['pops'],
-                    table_lables,
-                    tableClickHandler = clickTablePopDesires
-                    )
-                draw_table(
-                    "factionTable",
-                    data['factions'],
-                    faction_table_lables
-                    )
-            }
-            addTextBox(d,data)
-        }
-    });
-}
-
-function draw_planet(pdata){
-    // console.log(pdata)
-    draw_node("pSystem",
-            pdata["nodes"],
-            pdata["links"],
-            p_objectColors,
-            .003,
-            height,
-            width,
-            strokesFunc = s_objectStrokes) 
-    draw_table(
-        "planetsTable",
-        pdata["nodes"],
-        planet_table_lables,  
-        tableClickHandler = clickTablePlanet
-    )
-}
-
-
-function clickTablePopDesires(d){
-    $.ajax({
-        url: '/ajax/pop-desires',
-        type: 'get',
-        data: d,
-        dataType: 'json',
-        beforeSend: function () {
-            d3.selectAll('#peopledesires').remove()
-        },
-        success: function(data){
-            console.log(data)
-            if ("desires" in data){
-                draw_table(
-                    "peopledesires",
-                    data['desires'],
-                    desire_table_lables
-                    )
-            }
-        }
-    });
-}
