+++ app/templates/app/creation/genesis_view.html
@@ -20,8 +20,6 @@
             height = 480,
             defaultNodeSize = 10
         console.log(account)
-        {% include "app/charts/base_nodemap.js" %}
-        {% include "app/charts/base_table.js" %}
         {% include "app/creation/genesis.js" %}
     </script>
