+++ app/templates/app/menu_top.html
@@ -1,5 +1,5 @@
 <div>
-    Greetings, {{ user.username }}.
+    Greetings, {{ account.username }}.
 </div>
 <div class="menu">2D Menu
     <p>
@@ -21,4 +21,4 @@
     <p>
         <a class='menu button' href='/'>Quit</a>
     </p>
-</div>
+</div>
