+++ app/views/structures.py
@@ -1,0 +1,8 @@
+from django.contrib.auth import authenticate, login
+from django.contrib.auth.decorators import login_required
+from django.shortcuts import redirect, render
+
+from django.conf import settings
+
+
+
